make compressed dataset
