+++ yucca/pipeline/managers/dataloading/YuccaManager_Only1Modality.py
@@ -1,5 +1,5 @@
 from yucca.pipeline.managers.YuccaManager import YuccaManager
-from yucca.data_loading.alternative_datasets.YuccaDataset_1modality import YuccaTrainDataset_1modality
+from yucca.data.datasets.alternative_datasets.YuccaDataset_1modality import YuccaTrainDataset_1modality
 
 
 class YuccaManager_Only1Modality(YuccaManager):
