+++ yucca/pipeline/managers/dataloading/YuccaManager_compressed.py
@@ -1,0 +1,15 @@
+from yucca.pipeline.managers.YuccaManager import YuccaManager
+from yucca.data.datasets.YuccaCompressedDataset import YuccaCompressedTrainDataset
+
+
+class YuccaManager_compressed(YuccaManager):
+    def __init__(
+        self,
+        *args,
+        **kwargs,
+    ):
+        super().__init__(
+            *args,
+            **kwargs,
+        )
+        self.train_dataset_class = YuccaCompressedTrainDataset
